Refactor Resource
Move class method into instance, this allows simpler initialisation and
allows us to expose #url, #headers and #payload methods which can be
tested separately from each other.

Also delays the bearer token remote request so this will happen when the
data is needed instead of when the Resource is initialised.